+++ Mercury.Common/src/Mercury.Common/IEntity.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 
 namespace Mercury.Common
 {
@@ -7,5 +8,9 @@ namespace Mercury.Common
         Guid Id { get; set; }
         DateTimeOffset CreatedAt { get; set; }
         DateTimeOffset UpdatedAt { get; set; }
+
+        bool IsValid { get; set; }
+
+        List<string> Errors { get; set; }
     }
 }