+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Mercury
 
-# Mercury.Common v1.0.4.7
+# Mercury.Common v1.0.4.9
 # Mercury.Reservations v1.0.1.0
 
-## 2021-07-08
+## 2021-08-21
+1. Adds validation uniqueness title and Id for Rooms
+
+## 2021-08-07
 1. Adds xUnit for TDD
 2. Allows the user to create rooms
 3. Allows the user to get all rooms